move template filter and model to gpui namespace
